Extract _kometa_version.py from _legacy.py
